+++ src/Macros/Response/Created.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace Apiato\Core\Macros\ResponseTransformer;
+namespace Apiato\Core\Macros\Response;
 
 use Apiato\Core\Abstracts\Transformers\Transformer;
 use Illuminate\Http\JsonResponse;
-use Apiato\Core\Services\ResponseTransformer;
+use Apiato\Core\Services\Response;
 
 class Created {
     public function __invoke(): callable
@@ -14,7 +14,7 @@ class Created {
              * Returns a 201 Created response.
              */
             function (): JsonResponse {
-                /** @var ResponseTransformer $this */
+                /** @var Response $this */
                 if (is_null($this->getTransformer())) {
                     $this->transformWith(Transformer::empty());
                 }
