+++ src/Macros/Response/Ok.php
@@ -1,24 +1,24 @@
 <?php
 
-namespace Apiato\Core\Macros\ResponseTransformer;
+namespace Apiato\Core\Macros\Response;
 
 use Apiato\Core\Abstracts\Transformers\Transformer;
 use Illuminate\Http\JsonResponse;
-use Apiato\Core\Services\ResponseTransformer;
+use Apiato\Core\Services\Response;
 
-class Accepted {
+class Ok {
     public function __invoke(): callable
     {
         return
             /**
-             * Returns a 202 Accepted response.
+             * Returns a 200 OK response.
              */
             function (): JsonResponse {
-                /** @var ResponseTransformer $this */
+                /** @var Response $this */
                 if (is_null($this->getTransformer())) {
                     $this->transformWith(Transformer::empty());
                 }
-                return $this->respond(202);
+                return $this->respond(200);
             };
     }
 }
